analyze: We can now get min, max, average, and best bin on average
This allows for analysis of best-performing values allowing for rapid
optimization of HPL (as fast as HPL will finish, at least). For instance,
binning the The Grand Cluster NB test results by NB and using getBestBin
to get the fastest time on average lets us see that the optimal NB for
a single core run on The Grand Cluster is 216, with our assumption, 224,
coming close to the time achieved by 216.